requirements.txt
# Core Framework
streamlit>=1.30.0

# LlamaIndex and RAG
llama-index>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-ollama>=0.1.0
llama-index-vector-stores-postgres>=0.1.0

# Vector Database
psycopg2-binary>=2.9.9
pgvector>=0.2.0
sqlalchemy>=2.0.0
numpy>=1.24.0

# Document Processing
pymupdf>=1.23.0

# Reranking and Hybrid Search
sentence-transformers>=2.2.0
rank-bm25>=0.2.2

# MCP Server
mcp>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0  # For testing FastAPI endpoints
faker>=20.0.0  # For generating test data
